requirements.txt
torch
torch-geometric
rdkit-pypi
streamlit
scikit-learn
pandas